requirements.txt
asyncio
influxdb-client
python-kasa